requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
python-dotenv==1.0.1
google-generativeai==0.8.3
openai>=1.30.0
Pillow==11.0.0
